Move explore in APM into actions
